% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{build_cross_sigma_single}
\alias{build_cross_sigma_single}
\title{Cross-covariance for a single new site}
\usage{
build_cross_sigma_single(A, rho, d_vec, L, p = 6)
}
\arguments{
\item{A}{Lower triangular matrix (p x p).}

\item{rho}{Range parameters (p).}

\item{d_vec}{Vector of distances from new site to each of L observed sites.}

\item{L}{Number of observed sites.}

\item{p}{Number of parameters (default 6).}
}
\value{
Matrix of dimension (p x Lp).
}
\description{
Builds the p x (L*p) cross-covariance matrix between a single new site
and all L observed sites.
}
\keyword{internal}
