% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{build_sigma}
\alias{build_sigma}
\title{Build the coregionalization covariance matrix}
\usage{
build_sigma(A, rho, D, p = 6)
}
\arguments{
\item{A}{Lower triangular matrix (p x p).}

\item{rho}{Vector of p range parameters.}

\item{D}{Distance matrix (L x L).}

\item{p}{Number of parameters (default 6).}
}
\value{
Covariance matrix of dimension (L\emph{p x L}p).
}
\description{
Constructs the covariance matrix for the spatial random effects
following Eq. (7) in Russell et al. (2020). For p parameters and L sites,
Sigma = (I_L x A) sum_i (e_i e_i^T x Omega_i(rho)) (I_L x A)^T.
}
\details{
In our setting, p = 6 (3 NOAA + 3 ADCIRC GEV parameters).
}
