% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{compute_cross_distances}
\alias{compute_cross_distances}
\title{Cross-distance matrix between two sets of sites}
\usage{
compute_cross_distances(sites1, sites2)
}
\arguments{
\item{sites1}{Data frame with lon, lat (n1 rows).}

\item{sites2}{Data frame with lon, lat (n2 rows).}
}
\value{
Matrix of dimension (n1 x n2) of great-circle distances in km.
}
\description{
Cross-distance matrix between two sets of sites
}
