% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_model.R
\name{default_start}
\alias{default_start}
\title{Generate default starting values}
\usage{
default_start(stage1, dat, p = NULL)
}
\arguments{
\item{stage1}{Stage 1 fits.}

\item{dat}{Data object.}

\item{p}{Dimension (default 6).}
}
\value{
Named list with beta, A, rho.
}
\description{
Generate default starting values
}
\keyword{internal}
