% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{exp_cov_matrix}
\alias{exp_cov_matrix}
\title{Exponential covariance matrix}
\usage{
exp_cov_matrix(D, rho)
}
\arguments{
\item{D}{Distance matrix.}

\item{rho}{Range parameter (positive).}
}
\value{
Correlation matrix of same dimension as D.
}
\description{
Computes the correlation matrix Omega_i where
Omega_i(j,k) = exp(-D(j,k) / rho) for range parameter rho > 0.
}
