% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_exceedance_prob}
\alias{gev_exceedance_prob}
\title{GEV exceedance probability}
\usage{
gev_exceedance_prob(z0, mu, sigma, xi)
}
\arguments{
\item{z0}{Threshold value.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter (NOT log scale).}

\item{xi}{Shape parameter.}
}
\value{
Exceedance probability.
}
\description{
Computes P(Z > z0) for Z ~ GEV(mu, sigma, xi).
}
