% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_return_level}
\alias{gev_return_level}
\title{GEV quantile function (return level)}
\usage{
gev_return_level(mu, sigma, xi, r)
}
\arguments{
\item{mu}{Location parameter.}

\item{sigma}{Scale parameter (NOT log scale).}

\item{xi}{Shape parameter.}

\item{r}{Return period in years.}
}
\value{
The r-year return level.
}
\description{
Computes the r-year return level from GEV parameters.
}
