% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_levels.R
\name{rl_gradient}
\alias{rl_gradient}
\title{Gradient of return level w.r.t. (mu, log_sigma, xi)}
\usage{
rl_gradient(mu, log_sigma, xi, r)
}
\arguments{
\item{mu}{Location parameter.}

\item{log_sigma}{Log of scale parameter.}

\item{xi}{Shape parameter.}

\item{r}{Return period.}
}
\value{
Numeric vector of length 3.
}
\description{
For the delta method. Note: parameterized in terms of log(sigma),
not sigma, to match the kriging output.
}
\keyword{internal}
