% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{wendland_taper_matrix}
\alias{wendland_taper_matrix}
\title{Wendland 2 taper correlation matrix}
\usage{
wendland_taper_matrix(D, lambda)
}
\arguments{
\item{D}{Distance matrix.}

\item{lambda}{Taper range (same units as D). Correlations are zero beyond this distance.}
}
\value{
Sparse taper matrix (class "dgCMatrix" from Matrix package).
}
\description{
Constructs a sparse taper matrix based on the Wendland C4 function
with compact support at distance lambda.
}
