% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evfuse-package.R
\docType{package}
\name{evfuse-package}
\alias{evfuse}
\alias{evfuse-package}
\title{evfuse: Spatial Data Fusion for Extreme Value Analysis}
\description{
Two-stage frequentist framework for fusing sparse observations and dense simulations for spatial extreme value analysis. Fits generalized extreme value (GEV) distributions independently at each site, then models the site-level parameter estimates jointly via a linear model of coregionalization with a selection matrix that accommodates partial observations. Developed for U.S. coastal sea levels (White et al.) but applicable to any setting where annual maxima from multiple spatial data sources can be fused.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BrianNathanWhite/evfuse}
  \item \url{https://briannathanwhite.github.io/evfuse/}
  \item Report bugs at \url{https://github.com/BrianNathanWhite/evfuse/issues}
}

}
\author{
\strong{Maintainer}: Brian N. White \email{bnw@unc.edu}

Authors:
\itemize{
  \item Brian Blanton
  \item Rick Luettich
  \item Richard L. Smith
}

}
